+++ src/tasks/alacritty.rs
@@ -1,12 +1,10 @@
-use std::path::Path;
-
 use utils;
 
 pub fn sync() {
     println!("pkg: alacritty: syncing ...");
 
-    let src = utils::env::home_dir().join(Path::new(".dotfiles/config/alacritty.yml"));
-    let dest = utils::env::home_dir().join(Path::new(".config/alacritty/alacritty.yml"));
+    let src = utils::env::home_dir().join(".dotfiles/config/alacritty.yml");
+    let dest = utils::env::home_dir().join(".config/alacritty/alacritty.yml");
 
     utils::fs::symbolic_link_if_exists(&src, &dest);
 }