+++ src/tasks/dep.rs
@@ -1,5 +1,4 @@
 use std;
-use std::path::Path;
 
 use utils;
 use utils::github::{Asset, Release};
@@ -52,9 +51,9 @@ fn install_release_asset(release: Release) {
     println!("pkg: dep: installing ...");
 
     #[cfg(windows)]
-    let bin_path = utils::env::home_dir().join(Path::new(".local/bin/dep.exe"));
+    let bin_path = utils::env::home_dir().join(".local/bin/dep.exe");
     #[cfg(not(windows))]
-    let bin_path = utils::env::home_dir().join(Path::new(".local/bin/dep"));
+    let bin_path = utils::env::home_dir().join(".local/bin/dep");
 
     utils::github::download_release_asset(asset, &bin_path);
 }