+++ src/tasks/dotfiles.rs
@@ -1,11 +1,9 @@
-use std::path::Path;
-
 use utils;
 
 pub fn sync() {
     println!("pkg: dotfiles: syncing ...");
 
-    let target = utils::env::home_dir().join(Path::new(".dotfiles"));
+    let target = utils::env::home_dir().join(".dotfiles");
 
     if utils::git::has_git() && utils::git::path_is_git_repository(&target) {
         utils::git::pull(&target);