+++ src/tasks/psql.rs
@@ -1,5 +1,3 @@
-use std::path::Path;
-
 use utils;
 
 pub fn sync() {
@@ -9,8 +7,8 @@ pub fn sync() {
 
     println!("pkg: psql: syncing ...");
 
-    let src = utils::env::home_dir().join(Path::new(".dotfiles/config/psqlrc"));
-    let dest = utils::env::home_dir().join(Path::new(".psqlrc"));
+    let src = utils::env::home_dir().join(".dotfiles/config/psqlrc");
+    let dest = utils::env::home_dir().join(".psqlrc");
 
     utils::fs::symbolic_link_if_exists(&src, &dest);
 }