+++ src/tasks/skaffold.rs
@@ -1,5 +1,4 @@
 use std;
-use std::path::Path;
 
 use utils;
 use utils::github::{Asset, Release};
@@ -56,9 +55,9 @@ fn install_release_asset(release: Release) {
     println!("pkg: skaffold: installing ...");
 
     #[cfg(windows)]
-    let bin_path = utils::env::home_dir().join(Path::new(".local/bin/skaffold.exe"));
+    let bin_path = utils::env::home_dir().join(".local/bin/skaffold.exe");
     #[cfg(not(windows))]
-    let bin_path = utils::env::home_dir().join(Path::new(".local/bin/skaffold"));
+    let bin_path = utils::env::home_dir().join(".local/bin/skaffold");
 
     utils::github::download_release_asset(asset, &bin_path);
 }